+++ app/components/AppText/index.js
@@ -1,0 +1,2 @@
+import AppText from "./AppText";
+export default AppText;
